BEAUTYCAMERA/requirements.txt
streamlit
opencv-python-headless==4.7.0.72
numpy
Pillow